+++ package.json
@@ -30,7 +30,6 @@
   "homepage": "https://github.com/tnguyen14/notes#readme",
   "dependencies": {
     "@tridnguyen/auth": "^2.1.0",
-    "async": "^2.0.0",
     "bluebird": "^3.4.1",
     "body-parser": "^1.15.2",
     "cookie-session": "^2.0.0-alpha.1",
